requirements.txt
asgiref==3.7.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Deprecated==1.2.13
Django==4.2.1
djangorestframework==3.14.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
ogr==0.44.0
Pillow==9.5.0
postgis==1.0.4
psycopg==3.1.9
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
PyGithub==1.58.2
PyJWT==2.7.0
PyNaCl==1.5.0
python-gitlab==3.14.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-toolbelt==1.0.0
shortuuid==1.0.11
smmap==5.0.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
wrapt==1.15.0
